move some more files / remove root.module.css
