+++ packages/core/src/extensions/DraggableBlocks/DragMenuFactoryTypes.ts
@@ -1,4 +1,4 @@
-import { EditorElement, ElementFactory } from "../../EditorElement";
+import { EditorElement, ElementFactory } from "../../shared/EditorElement";
 
 export type AddBlockButtonParams = {
   addBlock: () => void;
