+++ packages/core/src/extensions/Hyperlinks/HyperlinkMenuFactoryTypes.ts
@@ -1,4 +1,4 @@
-import { EditorElement, ElementFactory } from "../../EditorElement";
+import { EditorElement, ElementFactory } from "../../shared/EditorElement";
 
 export type HyperlinkMenuParams = {
   url: string;
