+++ packages/core/src/shared/EditorElement.ts
